requirements.txt
starter_api
elasticsearch==2.3.0
jsonklog==0.15.0
fluent-logger==0.4.0
six==1.10.0